Avoid caching Exception when loading shrine and gate data
Shrine and Gate names are really just nice-to-haves. Before this commit,
if the data loading fails in an exception, the result would be captured
in a future, which would cause the same exception to be thrown everytime
it is accessed.

To fix this, we catch the exception and return an empty list in place of
actual loaded/parsed data. The end result is the app would lose the
ability to look up and display shrine and gate names. The program would
continue to function normally without said names.